+++ eutils/ebasearray.h
@@ -15,11 +15,11 @@ class ebasearray
    void load(const estr& file);
 
    virtual void serial(estr& data) const;
-   virtual int unserial(const estr& data,int i=0);
+   virtual size_t unserial(const estr& data,size_t i=0);
 
-   virtual unsigned int size() const=0;
+   virtual size_t size() const=0;
    virtual void addvar(evar& var)=0;
-   virtual evar getvar(int i) const=0;
+   virtual evar getvar(size_t i) const=0;
 };
 
 ostream& operator<<(ostream& stream,const ebasearray& arr);
