+++ eutils/ebasearrayof.h
@@ -11,13 +11,13 @@ class ebasearrayof
 {
  public:
    void serial(estr& data) const;
-   int unserial(const estr& data,int i=0);
+   size_t unserial(const estr& data,size_t i=0);
 
-   virtual unsigned int size() const=0;
+   virtual size_t size() const=0;
    virtual void addvar(evar& key,evar& var)=0;
    virtual evar getvarByKey(const evar& key) const=0;
-   virtual evar getvar(int i) const=0;
-   virtual evar getvarkey(int i) const=0;
+   virtual evar getvar(size_t i) const=0;
+   virtual evar getvarkey(size_t i) const=0;
 };
 
 ostream& operator<<(ostream& stream,const ebasearrayof& arr);
