+++ eutils/ebasicarray.h
@@ -26,9 +26,9 @@ evar ebasicarray<evar*>::getvar(int i) const;
 */
 
 template <class T>
-void ebasicarray<T>::init(int count)
+void ebasicarray<T>::init(size_t count)
 {
-  int i;
+  size_t i;
   clear();
   vector<T>::reserve(count);
   for (i=0; i<count; ++i)
@@ -36,9 +36,9 @@ void ebasicarray<T>::init(int count)
 }
 
 template <class T>
-void ebasicarray<T>::init(int count,const T& initval)
+void ebasicarray<T>::init(size_t count,const T& initval)
 {
-  int i;
+  size_t i;
   clear();
   vector<T>::reserve(count);
   for (i=0; i<count; ++i)
@@ -53,7 +53,7 @@ void ebasicarray<T>::serial(estr& data) const
 
   if (!getClasses().exists(typeid(T).name())) { lerror(estr("class not registered: ")+typeid(T).name()); return; }
 
-  int ilen=data.len();
+  size_t ilen=data.len();
 
   eclassBase *pClass=&getClasses().values(typeid(T).name());
   evarType<T> *pType=new evarType<T>((T&)at(0));
@@ -62,7 +62,7 @@ void ebasicarray<T>::serial(estr& data) const
 
   if (data.len()-ilen>0){ data.reserve(ilen+ (data.len()-ilen)*size()); }
 
-  int i;
+  size_t i;
   for (i=1; i<size(); ++i){
     pType->object=(T*)&at(i);
     pClass->fserial(pType,data);
@@ -79,7 +79,7 @@ void ebasicarray<T>::serial(estr& data) const
 
 
 template <class T>
-int ebasicarray<T>::unserial(const estr& data,int i)
+size_t ebasicarray<T>::unserial(const estr& data,size_t i)
 {
   clear();
   unsigned int count;
@@ -110,7 +110,7 @@ void ebasicarray<T>::addvar(evar& var)
   add(var.get<T>());
 }
 template <class T>
-evar ebasicarray<T>::getvar(int i) const
+evar ebasicarray<T>::getvar(size_t i) const
 {
   return(evar((T&)at(i)));
 }
@@ -118,7 +118,7 @@ evar ebasicarray<T>::getvar(int i) const
 template <>
 void ebasicarray<evar>::addvar(evar& var);
 template <>
-evar ebasicarray<evar>::getvar(int i) const;
+evar ebasicarray<evar>::getvar(size_t i) const;
 
 
 
@@ -126,7 +126,7 @@ template <class T>
 ostream &operator<<(ostream &stream,const ebasicarray<T> &var)
 {
   stream<<"{ "<<endl;
-  int i;
+  size_t i;
 
   if (var.size()){
     for (i=0; i<var.size()-1; ++i){
@@ -229,7 +229,7 @@ ebasicarray<T> &ebasicarray<T>::operator=(const ebasicarray<T> &arr)
 template <class T>
 ebasicarray<T> &ebasicarray<T>::operator+=(const ebasicarray<T> &arr)
 {
-  int i;
+  size_t i;
   for (i=0; i<arr.size(); ++i)
     push_back(arr[i]);
   return(*this);
@@ -239,7 +239,7 @@ template <class T>
 ebasicarray<T> ebasicarray<T>::operator[](const eintarray& arr) const
 {
   ebasicarray<T> res;
-  int i;
+  size_t i;
   for (i=0; i<arr.size(); ++i){
     if (arr[i]<0 || arr[i]>=size()) { ldwarn("element in array list out of bounds: "+estr(arr[i])); continue; }
     res.add(operator[](arr[i]));
@@ -250,7 +250,7 @@ ebasicarray<T> ebasicarray<T>::operator[](const eintarray& arr) const
 template <class T>
 ebasicarray<T> &ebasicarray<T>::operator-=(const ebasicarray<T> &arr)
 {
-  int ind,j;
+  size_t ind,j;
   for (j=0; j<arr.size(); ++j){
     if (-1 != ind = find(arr[j]))
       vector<T>::erase(vector<T>::begin()+ind);
@@ -293,7 +293,7 @@ ebasicarray<T> ebasicarray<T>::operator[](const ebasicarray<int> &iarr)
 */
 
 template <class T>
-ebasicarray<T> ebasicarray<T>::subset(int i,int l) const
+ebasicarray<T> ebasicarray<T>::subset(size_t i,size_t l) const
 {
   ebasicarray<T> tmpa;
  
@@ -309,7 +309,7 @@ ebasicarray<T> ebasicarray<T>::subset(int i,int l) const
 }
 
 template <class T>
-int ebasicarray<T>::find(const T &value,int i,bool (*match)(const T &a, const T &b)) const
+long ebasicarray<T>::find(const T &value,size_t i,bool (*match)(const T &a, const T &b)) const
 {
   if (i<0) i+=vector<T>::size();
 
@@ -341,7 +341,7 @@ ebasicarray<T> ebasicarray<T>::afindall(const T &value,int i,bool (*match)(const
 */
 
 template <class T>
-ebasicarray<T> ebasicarray<T>::afindall(const ebasicarray<T> &arr,int i,bool (*match)(const T &a, const T &b)) const
+ebasicarray<T> ebasicarray<T>::afindall(const ebasicarray<T> &arr,size_t i,bool (*match)(const T &a, const T &b)) const
 {
   ebasicarray<T> tmpai;
 
@@ -356,7 +356,7 @@ ebasicarray<T> ebasicarray<T>::afindall(const ebasicarray<T> &arr,int i,bool (*m
 }
 
 template <class T>
-eintarray ebasicarray<T>::findall(const T &value,int i,bool (*match)(const T &a, const T &b)) const
+eintarray ebasicarray<T>::findall(const T &value,size_t i,bool (*match)(const T &a, const T &b)) const
 {
   eintarray tmpai;
 
@@ -371,7 +371,7 @@ eintarray ebasicarray<T>::findall(const T &value,int i,bool (*match)(const T &a,
 }
 
 template <class T>
-eintarray ebasicarray<T>::findall(const ebasicarray<T> &arr,int i,bool (*match)(const T &a, const T &b)) const
+eintarray ebasicarray<T>::findall(const ebasicarray<T> &arr,size_t i,bool (*match)(const T &a, const T &b)) const
 {
   eintarray tmpai;
